test(terminal): Doing TTD on transaction feature
nfc_manager is the most popular plugin to implement nfc features on flutter
Looks like one of the contributors has a different linter ( or maybe not using one). This is why the change on main.dart are automated
NpcPay is a bad name since the technology used here if called NFC. And since that feature is supposed to handdle credit card paymenet , that explain the new name of the entity CreditCard
We simulate NfcManager with MockNfcManager thks to mockito.

Linked #49